refactor requestss

+++ dev/Workshop/Workshop/DomainLayer/MarketPackage/IMarketController.cs
@@ -42,7 +42,7 @@ namespace Workshop.DomainLayer.MarketPackage
 
         StoreDTO getStoreInfo(int userId, string user, int storeId);
 
-        List<ProductDTO> SearchProduct(int userId, string user, string keyWords, string catagory, double minPrice, double maxPrice, int productReview);
+        List<ProductDTO> SearchProduct(int userId, string user, string keyWords, string catagory, double minPrice, double maxPrice, double productReview);
 
         double BuyCart(int userId, string user, string address);
